enhanceNTweaks: fix out of bounds on can make recipe
when the list was toggled, the cursor was not resetting, this caused
an issue where the cursor stayed in an invalid position, desynchronizing
the UI display and the selected item (and also giving errors).

Resetting the cursor to the first item will prevent this kind of issue
as long as you have any item you can make.